requirements.txt
cysystemd==1.6.0
dbus-python==1.3.2
pycairo==1.25.1
PyGObject==3.46.0
systemd==0.17.1
